ReGoapNode changes to goal calculation, dropped "support" to false preconditions
I have changed the core of ReGoapNode, where the new goal is calculated,
this is quite big for old codes, also I have dropped the false
preconditions support (the planner basically was ignoring false
preconditions with the predicate in MissingState, removed this since
it's the same but force the library's users to remove any false
precondition).
Now the whole action precondition is added to goal and then the goal is
cleared with the world's state and the current action effects.